+++ src/main/java/br/com/zup/gerenciadorDePostagem/postagem/PostagemRepository.java
@@ -2,6 +2,6 @@ package br.com.zup.gerenciadorDePostagem.postagem;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostagemRepository extends CrudRepository<Postagem,Integer> {
+public interface PostagemRepository extends CrudRepository<Postagem,Long> {
 
 }
